# internal/generated/api/openapi.yaml
openapi: 3.1.1
info:
  description: This is a simple API to list URLs and port mappings from a Nomad cluster
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Molecule
  version: 1.0.0
servers:
- description: Production server
  url: https://api.molecule.dbyte.xyz
security:
- {}
paths:
...
  /v1/services/{service}/alloc-restart:
    parameters:
    - description: The name of the service
      explode: false
      in: path
      name: service
      required: true
      schema:
        example: molecule
        pattern: "^[a-z0-9-]+$"
        type: string
      style: simple
    post:
      operationId: restart_service_allocations
      parameters:
      - description: The name of the service
        explode: false
        in: path
        name: service
        required: true
        schema:
          example: molecule
          pattern: "^[a-z0-9-]+$"
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/getURLs_400_response"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/getURLs_400_response"
          description: Invalid request
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/getURLs_400_response"
          description: Internal server error
      summary: Restart all allocations of a service
components:
  schemas:
...
    getURLs_400_response:
      example:
        message: message
        status: success
      properties:
        status:
          enum:
          - success
          - error
          type: string
        message:
          type: string
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: X-API-Key
      type: apiKey
    BasicAuth:
      scheme: basic
      type: http